refactor: align feature packages with mvvm
