+++ app/src/main/java/com/techmarketplace/feature/auth/data/AuthRepository.kt
@@ -1,4 +1,4 @@
-package com.techmarketplace.data.repository
+package com.techmarketplace.feature.auth.data
 
 import android.content.Context
 import com.techmarketplace.core.errors.AuthError
