+++ app/src/main/java/com/techmarketplace/feature/auth/presentation/login/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.techmarketplace.presentation.auth.view
+package com.techmarketplace.feature.auth.presentation.login
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.layout.imePadding
